requirements.txt
torch
scipy
librosa
numpy
praat-parselmouth
torchaudio

